# chargeFinderApp/go-client-generated/api/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: https://petstore3.swagger.io/api/v3
tags:
- name: charging station
  description: Everything about your stations
  externalDocs:
    description: Find out more
    url: http://swagger.io
- name: stations
  description: Access to Petstore orders
  externalDocs:
    description: Find out more about our store
    url: http://swagger.io
- name: charging
  description: Operations about user
paths:
  /availableStations:
    get:
      tags:
      - charging station
      description: Multiple status values can be provided with comma separated strings
      operationId: GetAvailableStations
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChargingStation'
                x-content-type: application/json
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChargingStation'
        "400":
          description: Invalid status value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
  /occupiedStations:
    get:
      tags:
      - charging station
      operationId: OccupiedStations
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChargingStation'
                x-content-type: application/json
        "400":
          description: Invalid status value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
  /chargingStations:
    get:
      tags:
      - stations
      operationId: GetLists
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChargingStation'
                x-content-type: application/json
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChargingStation'
        "400":
          description: Invalid status value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
    post:
      tags:
      - stations
      summary: add charging station
      operationId: AddChargingStation
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStation'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStation'
  /startCharging:
    get:
      tags:
      - charging
      operationId: GetChargingStations
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StartCharging'
                x-content-type: application/json
        "400":
          description: Invalid status value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
    post:
      tags:
      - charging
      operationId: StartChargings
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartCharging'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartCharging'
components:
  schemas:
    ChargingStation:
      type: object
      properties:
        StationID:
          type: integer
          format: int64
        EnergyOutput:
          type: string
        Type:
          type: string
        Availability:
          type: boolean
          format: date-time
        AvailabilityTime:
          type: string
          example: approved
      example:
        Availability: true
        Type: Type
        EnergyOutput: EnergyOutput
        AvailabilityTime: approved
        StationID: 0
    StartCharging:
      type: object
      properties:
        StationId:
          type: integer
          format: int64
        VehicleBatteryCapacity:
          type: string
        CurrentVehicleCharge:
          type: string
        ChargingStartTime:
          type: string
      example:
        CurrentVehicleCharge: CurrentVehicleCharge
        ChargingStartTime: ChargingStartTime
        StationId: 0
        VehicleBatteryCapacity: VehicleBatteryCapacity
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header